Abstract execute from execute_fee
